+++ Challenges/Xorset/xorset.c
@@ -3,16 +3,16 @@
 #include <stdlib.h>
 
 typedef struct TestCase {
-    unsigned int n;
-    unsigned int k;
+    size_t n;
+    size_t k;
 } TestCase;
 
-TestCase **parseTests(TestCase **tests, unsigned int nTests, char **argv);
+TestCase **parseTests(TestCase **tests, size_t nTests, char **argv);
 
 int main(int argc, char **argv) {
     // TODO: Add error checks for argc != 2; argv[1] should point to input file.
 
-    unsigned int nTests = argc - 2;
+    size_t nTests = argc - 2;
     TestCase *tests = malloc(nTests * sizeof(TestCase));
     parseTests(&tests, nTests, argv);
 }
@@ -22,7 +22,7 @@ int main(int argc, char **argv) {
  * @param nTests    The number of test cases.
  * @param argv      TODO: Rename. | The string representation of test cases.
  **/
-TestCase **parseTests(TestCase **tests, unsigned int nTests, char **argv) {
+TestCase **parseTests(TestCase **tests, size_t nTests, char **argv) {
     for (int i = 0; i < nTests; ++i) {
         char *input = "";
         strcpy(input, argv[i]);
